requirements.txt
flask
google-genai
groq
psutil
pyautogui
requests
sounddevice
kokoro-onnx
keyboard